// apps/api/src/common/enums/roles.enum.ts
export enum Role {
  SUPER_ADMIN = 'SUPER_ADMIN',
  SYSTEM_ADMIN = 'SYSTEM_ADMIN',
  HOSPITAL_ADMIN = 'HOSPITAL_ADMIN',
  DOCTOR = 'DOCTOR',
  NURSE = 'NURSE',
  PHARMACIST = 'PHARMACIST',
  LAB_SCIENTIST = 'LAB_SCIENTIST',
  RADIOLOGIST = 'RADIOLOGIST',
  RECEPTIONIST = 'RECEPTIONIST',
  BILLING_OFFICER = 'BILLING_OFFICER',
  PATIENT = 'PATIENT',
  INSURANCE_PROVIDER = 'INSURANCE_PROVIDER',
  ENTERPRISE_CLIENT = 'ENTERPRISE_CLIENT',
}